requirements.txt
requests
lxml
beautifulsoup4
weaviate-client
openai
markdownify
neo4j
pandas
nltk
jsonschema
sentence-transformers